Status Check, aka statuschk.
Monitor various services and notify state changes via email or pager.

Written by Simon Lockhart for Bogons 2006-2008, this is the original
version which has since been developed and enhanced by the BBC, but
that version is not open source.

This version works but the number of available modules is limited.